+++ src/common/navigation/RootNavigator.tsx
@@ -1,7 +1,7 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
 import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
-import useAxiosConfig from '../hooks/useAxiosConfig';
+
 import useCachedResources from '../hooks/useCachedResources';
 import { RootStackParamList } from './types';
 
@@ -13,10 +13,10 @@ import AuthStackNavigator from '../../auth/navigation/AuthStackNavigator';
 // screens
 import SplashScreen from '../screens/SplashScreen';
 import NotFoundScreen from '../screens/NotFoundScreen';
+import useAxiosConfig from '@common/hooks/useAxiosConfig';
 export default function RootNavigator() {
   const Stack = createNativeStackNavigator<RootStackParamList>();
   const isLoadingComplete = useCachedResources();
-
   useAxiosConfig();
 
   // check tokens