+++ src/profile/hooks/useUserApi.ts
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 export function useUserApi() {
   const getUserByUsername = (username: string) => {
-    return axios.get(`user/${username}`);
+    return axios.get(`users/${username}`);
   };
   return { getUserByUsername };
 }